pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluent_mcp"
version = "0.1.0"
description = "Fluent MCP - A modern package for MCP servers"
readme = "README.md"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "mcp @ git+https://github.com/modelcontextprotocol/python-sdk.git@v1.4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=7.0.0",
    "pytest-cov>=4.0.0"
]

[project.scripts]
fluent-mcp = "fluent_mcp.cli:main"

[tool.setuptools]
packages = ["fluent_mcp"]

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
